Some small fixes for tests

+++ valen-game-greatcircle/src/main/java/be/twofold/valen/game/greatcircle/reader/md6mesh/Md6MeshReader.java
@@ -98,9 +98,9 @@ public final class Md6MeshReader implements AssetReader<Model, GreatCircleAsset>
                 .orElseThrow();
 
             // Just assume it's a byte buffer, because we read it as such
-            var array = ((ByteBuffer) joints.buffer()).array();
+            var array = ((ShortBuffer) joints.buffer()).array();
             for (var j = 0; j < array.length; j++) {
-                array[j] = lookup[Byte.toUnsignedInt(array[j]) + meshInfo.unknown2()];
+                array[j] = lookup[Short.toUnsignedInt(array[j]) + meshInfo.unknown2()];
             }
         }
     }
